MSP430: accept all interrupt numbers and don't emit alias
This works around compile error that interrupt vector number in the
interrupt() attribute is out of bounds, when building with TI MSP GCC.

Despite both GCC and Clang using the same linker script, for Clang, we
need to manually create a symbol and allocate it in the per-vector
section (named with a name that matches the names that the linker script
expects). Note that Clang creates __isr_* symbols automatically, but
these are only interpreted by the deprecated MSPGCC compiler/linker, not
by TI GCC for MSP. See more on this below.

Clang creates __isr_* symbols for MSPGCC (see lib/CodeGen/TargetInfo.cpp
MSP430TargetCodeGenInfo::SetTargetAttributes), these are not relevant
for TI GCC, but they are not entirely harmless. They cause a problem
because Clang generates __isr_N where N is divided by 2 (because MSPGCC
wants it that way, I guess). So, if we pass a vector number, the odd and
even vector numbers will generate the same symbol and therefore linking
will fail with a multiply-defined symbol error. As a workaround we pass
the doubled value.

There is one more problem: Clang rejects vector numbers that are odd or
above 30. So, this problem with the above problem form a catch-22
situation: if we pass N/2 then the atttribute is accepted, but we get
multiply-defined values; if we pass N, it's rejected; and if we pass
N*2, it's rejected. So, the workaround is to not emit the __isr symbol
at all, and to accept the full range of interrupt vector numbers.

Patch by Alexei Colin <ac@alexeicolin.com> published at
https://lists.llvm.org/pipermail/llvm-dev/2018-May/123597.html